+++ src/hls/__init__.py
@@ -1,3 +1,3 @@
 """Hook Line Sync."""
 
-__version__ = "0.8.29.2"
+__version__ = "0.8.29.3"